requirements.txt
appdirs==1.4.3
asgiref==3.2.3
beautifulsoup4==4.8.2
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
cssselect==1.1.0
Django==3.0.2
django-appconf==1.0.3
django-ckeditor==5.7.1
django-cleanup==4.0.0
django-image-cropping==1.3.0
django-js-asset==1.2.2
djangorestframework==3.10.3
easy-thumbnails==2.7
fake-useragent==0.1.11
gunicorn==19.9.0
idna==2.8
lxml==4.5.0
parse==1.14.0
pilkit==2.0
Pillow==6.2.1
pkg-resources==0.0.0
psycopg2==2.8.4
pyee==7.0.1
pyppeteer==0.0.25
pyquery==1.4.1
pytz==2019.3
requests==2.22.0
requests-html==0.10.0
six==1.14.0
soupsieve==1.9.5
sqlparse==0.3.0
tqdm==4.42.1
urllib3==1.25.8
w3lib==1.21.0
websockets==8.1
